master -> develop

+++ storage/replica.go
@@ -190,6 +190,9 @@ func (d *atomicRangeDesc) load() *roachpb.RangeDescriptor {
 // using a lock, the copy might be inconsistent.
 func (d *atomicRangeDesc) String() string {
 	inconsistentDesc := d.load()
+	if !inconsistentDesc.IsInitialized() {
+		return fmt.Sprintf("%d{-}", inconsistentDesc.RangeID)
+	}
 	return fmt.Sprintf("%d{%s-%s}",
 		inconsistentDesc.RangeID, inconsistentDesc.StartKey, inconsistentDesc.EndKey)
 }
@@ -345,12 +348,6 @@ type KeyRange interface {
 
 var _ KeyRange = &Replica{}
 
-type tickInfo struct {
-	Exists             bool
-	IsRangeLeaseHolder bool
-	IsRaftLeader       bool
-}
-
 // withRaftGroupLocked calls the supplied function with the (lazily
 // initialized) Raft group. It assumes that the Replica lock is held.
 func (r *Replica) withRaftGroupLocked(f func(r *raft.RawNode) error) error {
@@ -496,20 +493,20 @@ func (r *Replica) newReplicaInner(
 // require a lock and its output may not be atomic with other ongoing work in
 // the replica. This is done to prevent deadlocks in logging sites.
 func (r *Replica) String() string {
-	return fmt.Sprintf("%s %s", r.store, &r.rangeDesc)
+	return fmt.Sprintf("[n%d,s%d,r%s]", r.store.Ident.NodeID, r.store.Ident.StoreID, &r.rangeDesc)
 }
 
 // Destroy clears pending command queue by sending each pending
 // command an error and cleans up all data associated with this range.
-func (r *Replica) Destroy(origDesc roachpb.RangeDescriptor) error {
+func (r *Replica) Destroy(origDesc roachpb.RangeDescriptor, destroyData bool) error {
 	desc := r.Desc()
 	if repDesc, ok := desc.GetReplicaDescriptor(r.store.StoreID()); ok && repDesc.ReplicaID >= origDesc.NextReplicaID {
 		return errors.Errorf("cannot destroy replica %s; replica ID has changed (%s >= %s)",
 			r, repDesc.ReplicaID, origDesc.NextReplicaID)
 	}
 
-	// Clear the pending command queue.
 	r.mu.Lock()
+	// Clear the pending command queue.
 	for _, p := range r.mu.pendingCmds {
 		p.done <- roachpb.ResponseWithError{
 			Reply: &roachpb.BatchResponse{},
@@ -523,6 +520,9 @@ func (r *Replica) Destroy(origDesc roachpb.RangeDescriptor) error {
 		r.mu.replicaID, r.RangeID)
 	r.mu.Unlock()
 
+	if !destroyData {
+		return nil
+	}
 	return r.store.destroyReplicaData(desc)
 }
 
@@ -924,6 +924,10 @@ func (r *Replica) setLastVerificationTimestamp(timestamp hlc.Timestamp) error {
 func (r *Replica) RaftStatus() *raft.Status {
 	r.mu.Lock()
 	defer r.mu.Unlock()
+	return r.raftStatusLocked()
+}
+
+func (r *Replica) raftStatusLocked() *raft.Status {
 	if rg := r.mu.internalRaftGroup; rg != nil {
 		return rg.Status()
 	}
@@ -1624,24 +1628,6 @@ func defaultProposeRaftCommandLocked(r *Replica, p *pendingCmd) error {
 	})
 }
 
-func (r *Replica) isRangeLeaseHolderLocked() bool {
-	lease := r.mu.state.Lease
-	if lease == nil {
-		return false
-	}
-	timestamp := r.store.Clock().Now()
-	if lease.Covers(timestamp) {
-		if lease.OwnedBy(r.store.Ident.StoreID) {
-			return true
-		}
-	}
-	return false
-}
-
-func (r *Replica) isRaftLeaderLocked() bool {
-	return r.mu.leaderID == r.mu.replicaID
-}
-
 func (r *Replica) handleRaftReady() error {
 	ctx := context.TODO()
 	var hasReady bool
@@ -1651,7 +1637,6 @@ func (r *Replica) handleRaftReady() error {
 	lastIndex := r.mu.lastIndex // used for append below
 	raftLogSize := r.mu.raftLogSize
 	leaderID := r.mu.leaderID
-
 	err := r.withRaftGroupLocked(func(raftGroup *raft.RawNode) error {
 		if hasReady = raftGroup.HasReady(); hasReady {
 			rd = raftGroup.Ready()
@@ -1698,10 +1683,8 @@ func (r *Replica) handleRaftReady() error {
 			r.store.mu.Lock()
 			defer r.store.mu.Unlock()
 
-			if _, exists := r.store.mu.replicaPlaceholders[r.RangeID]; exists {
-				if err := r.store.removePlaceholderLocked(r.RangeID); err != nil {
-					return errors.Wrapf(err, "could not remove placeholder before applySnapshot")
-				}
+			if r.store.removePlaceholderLocked(r.RangeID) {
+				atomic.AddInt32(&r.store.counts.filledPlaceholders, 1)
 			}
 			if err := r.store.processRangeDescriptorUpdateLocked(r); err != nil {
 				return errors.Wrapf(err, "could not processRangeDescriptorUpdate after applySnapshot")
@@ -1843,20 +1826,16 @@ func (r *Replica) handleRaftReady() error {
 	})
 }
 
-// tick the Raft group, returning any error. tick() updates the tickInfo
-// struct.
-func (r *Replica) tick(info *tickInfo) error {
+// tick the Raft group, returning any error and true if the raft group exists
+// and false otherwise.
+func (r *Replica) tick() (bool, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
-
-	info.IsRaftLeader = r.isRaftLeaderLocked()
-	info.IsRangeLeaseHolder = r.isRangeLeaseHolderLocked()
 
 	// If the raft group is uninitialized, do not initialize raft groups on
 	// tick.
 	if r.mu.internalRaftGroup == nil {
-		info.Exists = false
-		return nil
+		return false, nil
 	}
 
 	r.mu.ticks++
@@ -1871,12 +1850,10 @@ func (r *Replica) tick(info *tickInfo) error {
 		// cycles.
 		if err := r.refreshPendingCmdsLocked(
 			reasonTicks, r.store.ctx.RaftElectionTimeoutTicks); err != nil {
-			info.Exists = true
-			return err
+			return true, err
 		}
 	}
-	info.Exists = true
-	return nil
+	return true, nil
 }
 
 // pendingCmdSlice sorts by increasing MaxLeaseIndex.
